server: extract ldap operations to a crate
