+++ crates/domain/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod deserialize;
 pub mod public_schema;
 pub mod requests;
 pub mod schema;
