+++ crates/ldap/src/core/error.rs
