--- server/src/infra/ldap/mod.rs
@@ -1,7 +1,0 @@
-pub mod compare;
-pub mod create;
-pub mod delete;
-pub mod handler;
-pub mod modify;
-pub mod password;
-pub mod search;
