+++ server/src/infra/mod.rs
@@ -6,13 +6,9 @@ pub mod db_cleaner;
 pub mod graphql;
 pub mod healthcheck;
 pub mod jwt_sql_tables;
-pub mod ldap;
 pub mod ldap_server;
 pub mod logging;
 pub mod mail;
 pub mod sql_backend_handler;
 pub mod tcp_backend_handler;
 pub mod tcp_server;
-
-#[cfg(test)]
-pub mod test_utils;
